Fix issue where Heroku build was failing.
Essentially, in a previous commit, I upgraded the version of Spring boot the app uses. However, Spring boot 2.5.0 creates a `-plain` jar file that doesn't contain a manifest file. This in turn causes the app to fail to start on Heroku since an executable jar needs a manifest file and Heroku attempts to run all jars in `build/libs` ('java -Dserver.port=$PORT $JAVA_OPTS -jar build/libs/*.jar').

The fix is just not to create the -plain.jar file.