Add modified stylex webpack plugin.
For some reason, the current stylex webpack plugin converts any file that uses stylex into a commonjs module. However, the rest of the files use ES modules which is causing some components to break. As such, I just modified the webpack plugin to not convert the files into CommonJS modules.